Refactor Value classes into separate files
Move value classes into a separate folder and into separate files.
